vault_v1/pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "vault_v1"
version = "0.1.0"
description = "Minimal content-addressable storage vault with proofs"
authors = [{name = "Vault Team"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.23",
    "sqlalchemy>=2.0",
    "alembic>=1.11",
    "blake3>=0.4",
    "python-multipart>=0.0.6",
    "click>=8.1",
    "pydantic-settings>=2.0",
    "boto3>=1.28",
    "jwcrypto>=1.5",
    "python-dotenv>=1.0",
    "requests>=2.31"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "httpx>=0.27",
    "pytest-asyncio>=0.21",
    "ruff>=0.1.6"
]

[tool.setuptools.packages.find]
where = ["vault"]

[project.scripts]
vault = "vault.cli.main:cli"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100

